Reorganize code into separate files for better readability
